moved factory to separate file + more proper application summon
