+++ app/src/main/java/com/yalta/viewmodel/MapViewModelFactory.kt
@@ -1,0 +1,14 @@
+package com.yalta.viewmodel
+
+import android.app.Application
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+
+class MapViewModelFactory(
+    private val application: Application
+) :
+    ViewModelProvider.Factory {
+    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+        return MapViewModel(application) as T
+    }
+}
